+++ Src/OSF/CAIRO-1160/SRC/cairo-traps-compositor.c
@@ -79,7 +79,7 @@ static void do_unaligned_box(void (*blt)(void * closure, int16 x, int16 y, int16
 			do_unaligned_row(blt, closure, b, tx, y2, 1, _cairo_fixed_fractional_part(b->p2.y));
 	}
 	else
-		do_unaligned_row(blt, closure, b, tx, y1, 1, b->p2.y - b->p1.y);
+		do_unaligned_row(blt, closure, b, tx, y1, 1, static_cast<uint16>(b->p2.y - b->p1.y));
 }
 
 struct blt_in {
